+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # SMSGateway Change Log
 
+## Version 1.1.0 (2023-03-21)
+
+### Enhancements
+* Enhanced gateway example
+
+## Version 1.0.5 (2022-10-17)
+
+### Enhancements
+* Enhanced gateway example
+
 ## Version 1.0.4 (2022-10-13)
 
 ### Enhancements